ChangeList: add conversion tables from CP=1250,1252 (Latin-2,1 for Windows ANSI) to UNICODE for C/C++ compiler to support Windows source code.
[git-p4: depot-paths = "//depot/openwatcom/": change = 33373]

+++ bld/cc/c/coptions.c
@@ -143,8 +143,8 @@ static struct
 } SwData;
 
 // local variables
-static int character_encoding = 0;
-static long unicode_CP = 0;
+static int      character_encoding = 0;
+static unsigned unicode_CP = 0;
 
 int EqualChar( int c )
 {
