+++ bld/cc/h/cvars.h
@@ -748,7 +748,7 @@ extern  void    StmtInit( void );               /* cstmt */
 
 extern  void    FreeLiteral(STRING_LITERAL *);  /* cstring */
 extern  STRING_LITERAL  *GetLiteral(void);      /* cstring */
-extern  void    LoadUnicodeTable(long);         /* cstring */
+extern  void    LoadUnicodeTable(unsigned);     /* cstring */
 extern  void    StringInit(void);               /* cstring */
 extern  TREEPTR StringLeaf(int);                /* cstring */
